### 2019-11-07 #1869 @whedon
[PRE REVIEW]: EarthPy: A Python package that makes it easier to explore and plot raster and vector data using open source Python tools
**Submitting author:** @lwasser (<a href="http://orcid.org/0000-0002-8177-6550">Leah Wasser</a>)
**Repository:** <a href="https://github.com/earthlab/earthpy" target ="_blank">https://github.com/earthlab/earthpy</a>
**Version:** 0.7.6
