### 2019-11-07 #1869 @arfon
@lwasser - excellent to see the first pyOpenSci submission come through!

A couple of quick things:

- Can you link here to the review on the pyOpenSci-side of things?
- It looks like the references aren't compiling properly. From a quick inspection, it looks like the citation keys you're using in your `paper.md` aren't matching the ones in the `paper.bib`. For example in your paper you're using `kelsey_jordahl_2019_3483425:2019` but in the bibtex it's `kelsey_jordahl_2019_3483425`.

Once you've fixed those, you can ask Whedon to recompile your paper with `@whedon generate pdf`
