### 2019-11-07 #1869 @arfon
Looks like the pyOpenSci review happened in https://github.com/pyOpenSci/software-review/issues/3
