### 2019-11-07 #1869 @lwasser
ok @arfon let me first fix the paper / citation issues. My apologies for being slow to link the pyopensci review!! but thank you for finding it. [this link that you found](https://github.com/pyOpenSci/software-review/issues/3) is the correct one.

As far as reviewers @HaoZeke said he would be willing to review here. he did the pyopensci review. @jsta could also be a good fit for this review!!
