### 2019-11-07 #1869 @lwasser
ok @arfon the citations look better now. i copied the rasterio citation and it's unusual. i had submitted a pr that was accepted to fix it but it looks like the year  is rendering as n.d. because the date in the citation is 2013-pre . Would you like for me to adjust that citation in some way? I can also submit another PR to rasterio if that is preferred. i think sean wanted the citation to represent that hte tool was started in 2013 and is still being maintained in the date. they (mapbox) do not use zenodo to track doi's for their tool.
