### 2019-11-10 #1869 @arfon
@lwasser - could you make sure you've cited all of the software that you mention in your article. For example, Matplotlib seems to be missing a citation here: https://matplotlib.org/3.1.1/citing.html
