### 2019-11-12 #1869 @lwasser
hi @arfon !
I added matplotlib and pysal to the citations! I can't seem to find citation info for [georasters](https://github.com/ozak/georasters). This package was just included as an example of other packages that allow for the use of raster data. How would you like for me to proceed with citing that package?

The pdf seemed to generate well that citation aside.

Also it looks like Joe is overtasked with reviews right now. should i mention other people? or how do you proceed with the review itself to get reviewers on board? thank you!!
