### 2019-11-12 #1869 @HaoZeke
As mentioned by @lwasser, I would like to confirm that I would be happy to take this on when required, which might be relevant as I also reviewed it for the [initial pyopensci review](https://github.com/pyOpenSci/software-review/issues/3). @arfon
