### 2019-11-12 #1869 @arfon
> Also it looks like Joe is overtasked with reviews right now. should i mention other people? or how do you proceed with the review itself to get reviewers on board? thank you!!

@lwasser - just to clarify, we won't do a second review here. As the pyOpenSci review is compatible with the JOSS review we simply fast-track the paper through here.
