### 2019-11-12 #1886 @arfon
Accepting as this submission was reviewed in https://github.com/pyOpenSci/software-review/issues/3
