### 2019-11-12 #1869 @arfon
> i see @arfon thank you for the clarification!! please let me know what you need from me!!

Please see my comment over in https://github.com/openjournals/joss-reviews/issues/1886#issuecomment-552691390 .
