### 2019-11-12 #1886 @lwasser
ok @arfon I'm working on this now!! waiting for CI to pass and then a few more author edits to align with our zenodo release!!
