### 2019-11-13 #1886 @lwasser
Hi @arfon
1. here is the earthpy release that aligns with the JOSS review: https://github.com/earthlab/earthpy/releases/tag/JOSS
2. here is the associated ZENODO release that contains the paper title in it https://zenodo.org/record/3540957
3. i've add an author that is on zenodo that was missing from our paper to the paper in our github rep here: https://github.com/earthlab/earthpy/blob/master/paper.md

Please let me know if you need anything else to complete the review.
